Refactor onboarding logic and navigation structure
- Introduced the OnboardingResetter component to manage onboarding state and installation markers.
- Updated App component to integrate OnboardingResetter within the ErrorBoundary for improved error handling.
- Removed the obsolete NavigationProvider from the desktop client, streamlining the navigation structure.
- Cleaned up unused parseUrlSearchString utility function to enhance codebase maintainability.
